+++ counterparty-lib/counterpartylib/test/bytespersigop_test.py
@@ -4,7 +4,7 @@ import tempfile
 import bitcoin as bitcoinlib
 
 from counterpartylib.lib import ledger
-from counterpartylib.lib.v1 import api
+from counterpartylib.lib.api import api_v1 as api
 from counterpartylib.test import util_test
 from counterpartylib.test.fixtures.params import ADDR
